aws_s3_to_gcp_gcs/source/requirements.txt
# Function dependencies, for example:
# package>=version
boto3>=1.9.33
flask>=1.0.2
google-cloud-storage==1.15.0